feat(core-utils): deprecate non-localized features
no breaking changes.
